pyproject.toml
[project]
name = "pucky"
version = "0.1.0"
description = "A tiny coding agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai",      # for LLM calls
    "pathspec>=0.11.0",  # for .gitignore files filtering
]

[project.scripts]
pucky = "src.main:main"

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pyright>=1.1.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["pucky"]

[tool.ruff.lint.per-file-ignores]
"src/utils.py" = ["W291"]  # Trailing whitespace is intentional in ASCII art

[tool.pyright]
include = ["src"]
exclude = ["**/__pycache__", "**/.*"]
pythonVersion = "3.10"
typeCheckingMode = "basic"
reportMissingImports = true
reportMissingTypeStubs = false
